+++ src/minecraft/net/minecraft/src/BlockSoulSand.java
@@ -7,7 +7,7 @@ public class BlockSoulSand extends Block {
 
 	public AxisAlignedBB getCollisionBoundingBoxFromPool(World world1, int i2, int i3, int i4) {
 		float f5 = 0.125F;
-		return AxisAlignedBB.getBoundingBoxFromPool((double)i2, (double)i3, (double)i4, (double)(i2 + 1), (double)((float)(i3 + 1) - f5), (double)(i4 + 1));
+		return AxisAlignedBB.getBoundingBoxFromPool(i2, i3, i4, i2 + 1, (float)(i3 + 1) - f5, i4 + 1);
 	}
 
 	public void onEntityCollidedWithBlock(World world1, int i2, int i3, int i4, Entity entity5) {
